server_get: return component attributes when requested
The query mods being prepared for loading server component attributes
was incorrect. This fixes the server query mods and queries for
components and its attributes separately.

LoadInstalledFirmware, LoadComponentStatus and the other LoadComponent*
methods are specific to the Components model. They are not direct relations of the
Server object and so when querying from the Server context those methods
would fail.